Created a  potlucks helpers with 2 functions
